Dodanie mechanizmu do wyszukiwania po tytule ksiazki

+++ src/main/java/pl/epam/robot/database/dao/GenericDAO.java
@@ -1,7 +1,6 @@
 package pl.epam.robot.database.dao;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.List;
 
 import org.hibernate.Query;
@@ -21,5 +20,5 @@ public interface GenericDAO<T, ID extends Serializable> {
 	 
 	    public List<?> findAll(Class<?> clazz);
 	 
-	    public T findByID(Class<?> clazz, BigDecimal id);
+	    public T findByID(Class<?> clazz, Integer id);
 }
